# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/cases/":{"get":{"tags":["Cases"],"summary":"Get Cases","operationId":"get_cases_cases__get","parameters":[{"name":"page_number","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Page Number"}},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Page Size"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseListResponse"}}}},"404":{"description":"Endpoint not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Cases"],"summary":"Create Case","operationId":"create_case_cases__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseResponse"}}}},"404":{"description":"Endpoint not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{case_id}":{"get":{"tags":["Cases"],"summary":"Get Case","operationId":"get_case_cases__case_id__get","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"integer","title":"Case Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseWithApplicant"}}}},"404":{"description":"Endpoint not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Cases"],"summary":"Update Case","operationId":"update_case_cases__case_id__put","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"integer","title":"Case Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseResponse"}}}},"404":{"description":"Endpoint not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Cases"],"summary":"Delete Case","operationId":"delete_case_cases__case_id__delete","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"integer","title":"Case Id"}}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Endpoint not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/applicants/":{"get":{"tags":["Applicants"],"summary":"Get Applicants","operationId":"get_applicants_applicants__get","parameters":[{"name":"page_number","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Page Number"}},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Page Size"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicantListResponse"}}}},"404":{"description":"Endpoint not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Applicants"],"summary":"Create Applicant","operationId":"create_applicant_applicants__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicantCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicantResponse"}}}},"404":{"description":"Endpoint not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/applicants/{applicant_id}":{"get":{"tags":["Applicants"],"summary":"Get Applicant","operationId":"get_applicant_applicants__applicant_id__get","parameters":[{"name":"applicant_id","in":"path","required":true,"schema":{"type":"integer","title":"Applicant Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicantResponse"}}}},"404":{"description":"Endpoint not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Applicants"],"summary":"Update Applicant","operationId":"update_applicant_applicants__applicant_id__put","parameters":[{"name":"applicant_id","in":"path","required":true,"schema":{"type":"integer","title":"Applicant Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicantUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicantResponse"}}}},"404":{"description":"Endpoint not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Applicants"],"summary":"Delete Applicant","operationId":"delete_applicant_applicants__applicant_id__delete","parameters":[{"name":"applicant_id","in":"path","required":true,"schema":{"type":"integer","title":"Applicant Id"}}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Endpoint not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/":{"get":{"tags":["Users"],"summary":"Get Items","operationId":"get_items_users__get","parameters":[{"name":"page_number","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Page Number"}},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Page Size"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserListResponse"}}}},"404":{"description":"Endpoint not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Users"],"summary":"Create Item","operationId":"create_item_users__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"404":{"description":"Endpoint not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{id}":{"get":{"tags":["Users"],"summary":"Get Item","operationId":"get_item_users__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"404":{"description":"Endpoint not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Users"],"summary":"Update Item","operationId":"update_item_users__id__put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"404":{"description":"Endpoint not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Users"],"summary":"Delete Item","operationId":"delete_item_users__id__delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Endpoint not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/current-user":{"get":{"tags":["Admin"],"summary":"Get Current User","operationId":"get_current_user_admin_current_user_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/health/":{"get":{"tags":["Health"],"summary":"Get Health","operationId":"get_health_health__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ApplicantCreate":{"properties":{"first_name":{"type":"string","maxLength":50,"minLength":1,"title":"First Name"},"last_name":{"type":"string","maxLength":50,"minLength":1,"title":"Last Name"},"middle_name":{"anyOf":[{"type":"string","maxLength":50,"minLength":1},{"type":"null"}],"title":"Middle Name"},"gender":{"type":"string","maxLength":20,"minLength":1,"title":"Gender"},"date_of_birth":{"type":"string","format":"date","title":"Date Of Birth"},"ssn":{"type":"string","maxLength":11,"minLength":9,"title":"Ssn"},"email":{"anyOf":[{"type":"string","maxLength":254},{"type":"null"}],"title":"Email"},"home_phone":{"anyOf":[{"type":"string","maxLength":20},{"type":"null"}],"title":"Home Phone"},"mobile_phone":{"anyOf":[{"type":"string","maxLength":20},{"type":"null"}],"title":"Mobile Phone"},"address":{"anyOf":[{"type":"string","maxLength":200},{"type":"null"}],"title":"Address"},"city":{"anyOf":[{"type":"string","maxLength":100},{"type":"null"}],"title":"City"},"state":{"anyOf":[{"type":"string","maxLength":50},{"type":"null"}],"title":"State"},"zip":{"anyOf":[{"type":"string","maxLength":10},{"type":"null"}],"title":"Zip"},"country":{"type":"string","maxLength":100,"title":"Country","default":"USA"}},"type":"object","required":["first_name","last_name","gender","date_of_birth","ssn"],"title":"ApplicantCreate"},"ApplicantListResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/ApplicantResponse"},"type":"array","title":"Items"},"item_count":{"type":"integer","title":"Item Count","default":0},"page_count":{"type":"integer","title":"Page Count","default":0},"prev_page":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Prev Page"},"next_page":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Next Page"}},"type":"object","required":["items"],"title":"ApplicantListResponse"},"ApplicantResponse":{"properties":{"first_name":{"type":"string","maxLength":50,"minLength":1,"title":"First Name"},"last_name":{"type":"string","maxLength":50,"minLength":1,"title":"Last Name"},"middle_name":{"anyOf":[{"type":"string","maxLength":50,"minLength":1},{"type":"null"}],"title":"Middle Name"},"gender":{"type":"string","maxLength":20,"minLength":1,"title":"Gender"},"date_of_birth":{"type":"string","format":"date","title":"Date Of Birth"},"ssn":{"type":"string","maxLength":11,"minLength":9,"title":"Ssn"},"email":{"anyOf":[{"type":"string","maxLength":254},{"type":"null"}],"title":"Email"},"home_phone":{"anyOf":[{"type":"string","maxLength":20},{"type":"null"}],"title":"Home Phone"},"mobile_phone":{"anyOf":[{"type":"string","maxLength":20},{"type":"null"}],"title":"Mobile Phone"},"address":{"anyOf":[{"type":"string","maxLength":200},{"type":"null"}],"title":"Address"},"city":{"anyOf":[{"type":"string","maxLength":100},{"type":"null"}],"title":"City"},"state":{"anyOf":[{"type":"string","maxLength":50},{"type":"null"}],"title":"State"},"zip":{"anyOf":[{"type":"string","maxLength":10},{"type":"null"}],"title":"Zip"},"country":{"type":"string","maxLength":100,"title":"Country","default":"USA"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["first_name","last_name","gender","date_of_birth","ssn","id","created_at","updated_at"],"title":"ApplicantResponse"},"ApplicantUpdate":{"properties":{"first_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"First Name"},"last_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Name"},"middle_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Middle Name"},"gender":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Gender"},"date_of_birth":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date Of Birth"},"ssn":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ssn"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"home_phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Home Phone"},"mobile_phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mobile Phone"},"address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"},"zip":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Zip"},"country":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Country"}},"type":"object","title":"ApplicantUpdate"},"CaseCreate":{"properties":{"status":{"type":"string","enum":["Not Started","In Progress","Approved","Denied"],"title":"Status"},"assigned_to":{"anyOf":[{"type":"string","maxLength":255,"minLength":1},{"type":"null"}],"title":"Assigned To"},"applicant_id":{"type":"integer","title":"Applicant Id"}},"type":"object","required":["status","applicant_id"],"title":"CaseCreate"},"CaseListResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/CaseWithApplicant"},"type":"array","title":"Items"},"item_count":{"type":"integer","title":"Item Count"},"page_count":{"type":"integer","title":"Page Count"},"prev_page":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Prev Page"},"next_page":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Next Page"}},"type":"object","required":["items","item_count","page_count"],"title":"CaseListResponse"},"CaseResponse":{"properties":{"status":{"type":"string","enum":["Not Started","In Progress","Approved","Denied"],"title":"Status"},"assigned_to":{"anyOf":[{"type":"string","maxLength":255,"minLength":1},{"type":"null"}],"title":"Assigned To"},"id":{"type":"integer","title":"Id"},"applicant_id":{"type":"integer","title":"Applicant Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["status","id","applicant_id","created_at","updated_at"],"title":"CaseResponse"},"CaseUpdate":{"properties":{"status":{"anyOf":[{"type":"string","enum":["Not Started","In Progress","Approved","Denied"]},{"type":"null"}],"title":"Status"},"assigned_to":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Assigned To"}},"type":"object","title":"CaseUpdate"},"CaseWithApplicant":{"properties":{"status":{"type":"string","enum":["Not Started","In Progress","Approved","Denied"],"title":"Status"},"assigned_to":{"anyOf":[{"type":"string","maxLength":255,"minLength":1},{"type":"null"}],"title":"Assigned To"},"id":{"type":"integer","title":"Id"},"applicant_id":{"type":"integer","title":"Applicant Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"applicant":{"anyOf":[{"$ref":"#/components/schemas/ApplicantResponse"},{"type":"null"}]}},"type":"object","required":["status","id","applicant_id","created_at","updated_at"],"title":"CaseWithApplicant"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"UserCreate":{"properties":{"user_id":{"type":"string","maxLength":100,"minLength":1,"title":"User Id"},"first_name":{"type":"string","maxLength":100,"minLength":1,"title":"First Name"},"last_name":{"type":"string","maxLength":100,"minLength":1,"title":"Last Name"},"display_name":{"type":"string","maxLength":200,"minLength":1,"title":"Display Name"},"email":{"type":"string","maxLength":254,"minLength":1,"title":"Email"},"is_active":{"type":"boolean","title":"Is Active","default":true},"created_by":{"type":"string","maxLength":100,"minLength":1,"title":"Created By"},"modified_by":{"type":"string","maxLength":100,"minLength":1,"title":"Modified By"},"hashed_password":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Hashed Password"}},"type":"object","required":["user_id","first_name","last_name","display_name","email","created_by","modified_by"],"title":"UserCreate"},"UserListResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/UserResponse"},"type":"array","title":"Items"},"item_count":{"type":"integer","title":"Item Count","default":0},"page_count":{"type":"integer","title":"Page Count","default":0},"prev_page":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Prev Page"},"next_page":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Next Page"}},"type":"object","required":["items"],"title":"UserListResponse"},"UserResponse":{"properties":{"user_id":{"type":"string","maxLength":100,"minLength":1,"title":"User Id"},"first_name":{"type":"string","maxLength":100,"minLength":1,"title":"First Name"},"last_name":{"type":"string","maxLength":100,"minLength":1,"title":"Last Name"},"display_name":{"type":"string","maxLength":200,"minLength":1,"title":"Display Name"},"email":{"type":"string","maxLength":254,"minLength":1,"title":"Email"},"is_active":{"type":"boolean","title":"Is Active","default":true},"created_by":{"type":"string","maxLength":100,"minLength":1,"title":"Created By"},"modified_by":{"type":"string","maxLength":100,"minLength":1,"title":"Modified By"},"id":{"type":"integer","title":"Id"},"created":{"type":"string","format":"date-time","title":"Created"},"modified":{"type":"string","format":"date-time","title":"Modified"}},"type":"object","required":["user_id","first_name","last_name","display_name","email","created_by","modified_by","id","created","modified"],"title":"UserResponse"},"UserUpdate":{"properties":{"user_id":{"anyOf":[{"type":"string","maxLength":100,"minLength":1},{"type":"null"}],"title":"User Id"},"first_name":{"anyOf":[{"type":"string","maxLength":100,"minLength":1},{"type":"null"}],"title":"First Name"},"last_name":{"anyOf":[{"type":"string","maxLength":100,"minLength":1},{"type":"null"}],"title":"Last Name"},"display_name":{"anyOf":[{"type":"string","maxLength":200,"minLength":1},{"type":"null"}],"title":"Display Name"},"email":{"anyOf":[{"type":"string","maxLength":254,"minLength":1},{"type":"null"}],"title":"Email"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active"},"modified_by":{"anyOf":[{"type":"string","maxLength":100,"minLength":1},{"type":"null"}],"title":"Modified By"}},"type":"object","title":"UserUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}